requirements.txt
django>=4.0.0,<4.1.0
psycopg2
djangorestframework
djangorestframework-gis
pyyaml
requests
django-cors-headers
django-filter
prettyconf